requirements.txt
playsound
RPi.GPIO